Created KeepMeActive
This script is used to keep the current Windows user active. It will type a "." value every 60 seconds to ensure no timeouts occur. There are other approaches to this that may be more beneficial, such as using less-common keys instead of the ".", and ones that may not enter a value. For example, hitting the CTRL key every 5 minutes may be less intrusive.