Created ServerBootTime
This script can be used to reference whether a server restarted unexpectedly. Upon Windows Server startup, a task can be created to run this script to send out an email with the last boot time that notifies an individual or team that a server has restarted unexpectedly. Of course, the script does not tell whether a server is currently disconnected or offline.